+++ src/components/common/buttons/WatchlistButton.tsx
@@ -1,10 +1,10 @@
 import { ImSpinner2 } from "react-icons/im";
-import { trpc } from "../../utils/trpc";
-import { IThemeColor } from "../watchButton/BaseWatchButton";
 import ReactTooltip from "react-tooltip";
 import { toast } from "react-toastify";
 import { IoIosAdd, IoIosRemove, IoMdInformation } from "react-icons/io";
 import { BsBookmarkCheck, BsFillBookmarkDashFill } from "react-icons/bs";
+import { IThemeColor } from "../../watchButton/BaseWatchButton";
+import { trpc } from "../../../utils/trpc";
 
 interface IWatchlistButtonProps {
   themeColor: IThemeColor;
