+++ src/components/dashboard/UpNext.tsx
@@ -1,8 +1,8 @@
 import Image from "next/image";
 import Link from "next/link";
 import { PosterImage } from "../../utils/generateImages";
-import LoadingPageComponents from "./LoadingPageComponents";
-import { PosterGrid } from "./PosterGrid";
+import LoadingPageComponents from "../common/LoadingPageComponents";
+import { PosterGrid } from "../common/PosterGrid";
 import { AnimatePresence, motion } from "framer-motion";
 import { AiOutlineCheckCircle } from "react-icons/ai";
 import { trpc } from "../../utils/trpc";
