+++ src/components/pageBlocks/ContentPoster.tsx
@@ -5,8 +5,8 @@ import { useEffect, useState } from "react";
 import { AiFillCheckCircle } from "react-icons/ai";
 import ConditionalLink from "../../utils/ConditionalLink";
 import { PosterImage } from "../../utils/generateImages";
-import ReviewButton from "../common/ReviewButton";
-import WatchlistButton from "../common/WatchlistButton";
+import ReviewButton from "../common/buttons/ReviewButton";
+import WatchlistButton from "../common/buttons/WatchlistButton";
 import { IThemeColor } from "../watchButton/BaseWatchButton";
 import EpisodeWatchButton from "../watchButton/EpisodeWatchButton";
 import MovieWatchButton from "../watchButton/MovieWatchButton";
