+++ src/components/reviews/Review.tsx
@@ -1,9 +1,9 @@
 import Image from "next/image";
 import Link from "next/link";
 import { PosterImage } from "../../utils/generateImages";
-import ImageWithFallback from "./ImageWithFallback";
+import ImageWithFallback from "../common/ImageWithFallback";
 
-interface IFriendReview {
+interface IReview {
   content: string;
   created: Date;
   item: {
@@ -19,7 +19,7 @@ interface IFriendReview {
   hideImage?: boolean;
 }
 
-const FriendReview = ({ content, created, item, friend, hideImage }: IFriendReview) => {
+const Review = ({ content, created, item, friend, hideImage }: IReview) => {
   return (
     <div className="flex items-center w-full gap-2 mb-4">
       {!hideImage && (
@@ -61,7 +61,7 @@ const FriendReview = ({ content, created, item, friend, hideImage }: IFriendRevi
   );
 };
 
-export const LoadingFriendReview = () => (
+export const LoadingReview = () => (
   <div className="flex items-center w-full gap-2 mb-4">
     <div className="flex items-center w-full gap-2 mb-4">
       <div className="w-[100px] h-[150px] animate-pulse bg-[#343434]" />
@@ -78,4 +78,4 @@ export const LoadingFriendReview = () => (
   </div>
 );
 
-export default FriendReview;
+export default Review;
